Update RootContainer iOS implementation
Apply backgroundColor to SwiftUI implementation, too
