+++ composeApp/src/iosMain/kotlin/codes/chrishorner/personasns/WindowColorState.kt
@@ -2,13 +2,24 @@
 
 package codes.chrishorner.personasns
 
+import androidx.compose.ui.graphics.Color
+import codes.chrishorner.personasns.interop.RootContainerColor
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.map
 import platform.UIKit.UIColor
 
+private fun Color.toUIColor() = UIColor(
+    red = this.red.toDouble(),
+    green = this.green.toDouble(),
+    blue = this.blue.toDouble(),
+    alpha = this.alpha.toDouble(),
+)
+
 object WindowColorState {
-    internal val _statusBarColor = MutableStateFlow(UIColor.clearColor)
-    val statusBarColor = _statusBarColor.asStateFlow()
-    internal val _navigationBarColor = MutableStateFlow(UIColor.clearColor)
-    val navigationBarColor = _navigationBarColor.asStateFlow()
+    internal val rootContainerColorState =
+        MutableStateFlow(RootContainerColor(backgroundColor = Color.Transparent))
+
+    val statusBarColor = rootContainerColorState.map { it.statusBarColor.toUIColor() }
+    val navigationBarColor = rootContainerColorState.map { it.navigationBarColor.toUIColor() }
+    val backgroundColor = rootContainerColorState.map { it.backgroundColor.toUIColor() }
 }
