+++ composeApp/src/iosMain/kotlin/codes/chrishorner/personasns/interop/RootContainer.ios.kt
@@ -1,18 +1,14 @@
 package codes.chrishorner.personasns.interop
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.Modifier
 import codes.chrishorner.personasns.WindowColorState
-import platform.UIKit.UIColor
 import platform.UIKit.UIViewController
-
-private fun Color.toUIColor() = UIColor(
-    red = this.red.toDouble(),
-    green = this.green.toDouble(),
-    blue = this.blue.toDouble(),
-    alpha = this.alpha.toDouble(),
-)
 
 @Suppress("unused", "FunctionName")
 fun IosRootContainerEntryPoint(
@@ -21,15 +17,13 @@ fun IosRootContainerEntryPoint(
 
 @Composable
 actual fun RootContainer(
-    statusBarColor: Color,
-    navigationBarColor: Color,
-    content: @Composable () -> Unit
+    colors: RootContainerColor,
+    content: @Composable (BoxScope.() -> Unit)
 ) {
     // Create Compose view that wrap SwiftUI view that also contains Compose view
     SideEffect {
-        WindowColorState._statusBarColor.value = statusBarColor.toUIColor()
-        WindowColorState._navigationBarColor.value = navigationBarColor.toUIColor()
+        WindowColorState.rootContainerColorState.value = colors
     }
 
-    content()
+    Box(modifier = Modifier.fillMaxSize().background(colors.backgroundColor), content = content)
 }
